refactor: 할 일에서 기간 설정 컴포넌트 분리
